restructure and optimizations
