+++ Packet.js
@@ -1,0 +1,14 @@
+module.exports = {
+  'ESP3Packet': require('./ESP3/ESP3Packet.js'),
+  'RadioERP1': require('./ESP3/PacketTypes/RadioERP1.js'),
+  'Response': require('./ESP3/PacketTypes/Response'),
+  'RadioSubTel': require('./ESP3/PacketTypes/RadioSubTel'),
+  'Event': require('./ESP3/PacketTypes/Event'),
+  'CommonCommand': require('./ESP3/PacketTypes/CommonCommand'),
+  'SmartAckCommand': require('./ESP3/PacketTypes/SmartAckCommand'),
+  'RemoteManCommand': require('./ESP3/PacketTypes/RemoteManCommand'),
+  'RadioMessage': require('./ESP3/PacketTypes/RadioMessage'),
+  'RadioERP2': require('./ESP3/PacketTypes/RadioERP2'),
+  'Radio802': require('./ESP3/PacketTypes/Radio802'),
+  'Command24': require('./ESP3/PacketTypes/Command24')
+}
